requirements.txt
flask
numpy
pillow
tensorflow
keras